Correct nodepool max-servers in rax-dfw
* .../templates/nodepool/nodepool.yaml.erb: Since we still have a
handful of long-lived static slaves for sensitive jobs (proposal,
salt-trigger, mirrorXX) along with a couple dev slaves (precise,
centos6) and the remaining four precise-py3k slaves, the effective
capacity for nodepool in rax-dfw is 181. Cut this by a few more so
we have room for template servers during image updates.

Change-Id: I7cdfdba0665fff822fff77c31fa20a6585da411d