Stop running Grenade on Havana changes early
* modules/openstack_project/files/zuul/layout.yaml: Although the
stable/grizzly branch is not scheduled for EOL until March 6, we
are no longer testing upgrades from it to Havana proposed changes.
Recent non-backwards-compatible changes in python-swiftclient make
testing Swift with Grizzly more challenging, and nobody has time to
invest in fixing this for upgrade tests in the next week before
Grizzly is officially EOL. Instead stop caring about it so we can
merge some pending Havana fixes and move on.

Change-Id: I58a37a8ec7ac9ff302c807a18864421acb54005e