Don't ensure elasticsearch service is running
The package maintainer scripts start elasticsearch when it gets
installed, but it seems to return control before the pidfile is
created which allows puppet to race for it and ultimately launch a
second daemon resulting in an OOM condition. It's configured to
start at boot anyway, so the puppet ensure running isn't buying us
much.

Change-Id: I1b374f90c198b9cb73d412fe63b439623f434d9a