Fix update java alternatives on redhat systems
puppet was throwing the following error..
  nodepool.image.build.rax-iad.bare-centos6: Could not find dependent
  Exec[update-java-alternatives] for Package[ant] at |
  /root/config/modules/jenkins/manifests/slave.pp:86

because update-java-alternatives was only defined for debian, not for redhat systems.
This change adds java alternatives update to redhat.  Needed to change
'java-1.7.0-openjdk' to 'java-1.7.0-openjdk-devel' because 'java-1.7.0-openjdk' does
not install the javac compiler.

We are setting alternatives for debian and centos but not for fedora because the java 6
packages are not available on fedora.

Change-Id: I04b3e8292d3f31cd56ada9be99da8a234e13585a