Add in non-voting sample-config check
Until the sample-config is generated with documenation and during
packaging / setup time, there is a non-voting check job to indicate
the sample config in the git repository is out of date.

This has been enabled for Keystone, Nova, Heat, and Ceilometer (the
projects with the supported scripts from oslo to perform the check
and generate a sample config).

Change-Id: Ied16c842dbdd61fa8bc474c03e15370a52197f1a