Add a job to check IRC channel access
Add an IRC channel config file with some stub access information
for a few channels.  A new tool can read this file and verify
that openstackinfra has founder access on all the channels listed
in the file.

Run a job on changes to that file to verify that access is correct.

Later, other tools (like statusbot or gerritbot) may read that
file directly to determine which channels to join.  Or we can at
least add checks that all channels referenced by another tool
(like eavesdrop) are also listed in this file.

Related-Bug: 1190296
Change-Id: I38440e745402af5bbc3f0d0cc04a150c0a4bb47c