Limit nodepool nodes to 8GB of RAM
Limit the nodepool nodes to 8GB of RAM regardless of the flavor used.
This allows us to use arbitrary flavors that may have the CPU resources
that we require without risk that we become dependent on the vast
amounts of RAM that typically come with more CPU.

While we are at it set grub timeout to 0 saving 5 seconds of boottime
for every server we boot.

Change-Id: Ia38341239e0189c84a21a4f58dd7ca637490a27b