enable heat-slow in the integrated-gate
this would make heat-slow part of the integrated gate, it should be
a co-gating job because heat pushes on all the rest of the stack in
OpenStack, and tests bringing up non trivial guests (not cirros).

We remove the job from havana because it is not working there, and
better to not hold up enforcing this on icehouse.

Change-Id: Ia03840f7add545c0b11792cc2e7979415af4f15a