Put common global locations in a file of variables
There are some things that are referenced over and over again in
build scripts. Some of them might be interesting to be able to
override on a per-cloud-region location, such as pypi mirror. Others
are just things that are copied over and over and could be set once.

Put these things into two files - one in shell script format, one
in yaml format, so that scripts and programs can easily access the
data.

Change-Id: I54acf2b23d4a46c3b96593c42e57238b7e4c4beb