Temporarily stop using rax for py3k-precise nodes
* modules/openstack_project/templates/nodepool/nodepool.yaml.erb:
The py3k-precise nodes in rax regions are exhibiting an odd behavior
when using pip to upgrade setuptools along with other requirements
inside a pypy-based virtualenv. This is causing most pypy unit test
jobs run there to fail. Temporarily cease building new nodes of this
type in those locations for now, which will limit jobs to running on
the ones we've seen work successfully. This can be reverted once the
underlying discrepancy is determined and resolved.

Change-Id: I38156258a7a9b42cf536725fbaf75c7ddfa7bbd7