run link-logs builder first
The builders are run sequentially and short circuited. So if a job fails
we used to loose the link to the logs, but a failed job is the main
reason why someone would try to access the jenkins console log directly.

Confirmed this patch hits all cases of link-logs using 'git grep -B1 link-logs'

Change-Id: I74bc87b706c13624a1b34d0a0335c3a167ff8e34