Remove unneeded apt sources from nodepool nodes.
We need additional apt sources on nodepool nodes to install things like
puppet, pypy, and python3. But once these packages are installed we
don't need them anymore. Any test that does an apt-get update will hit
these apt repos despite not needing to. Remove the additional sources
after we are done with them to be nicer to them.

Fixes-Bug: 1253774
Change-Id: I3c5bd7e544076f26b224553bdea6134d39fc8bba