Add experimental virtual-ironic test
Add a test to set the necessary ENV flags and prepare the devstack
environment to test Ironic with "fake" bare metal VMs, and cause
devstack to configure Nova to use the "ironic" virt driver.

Create both voting and non-voting jobs for this test, and add them to
the appropriate pipelines. Make them all non-voting initially, and we'll
switch on the voting ones once things pass a few times.

Change-Id: I22abdaaab7fcd9c636f12dd53b7bdf43acbcc9cf