Fix the virtual-ironic job's export list
The job was trying to export ENV vars to devstack, but these aren't
passed all the way down. This patch depends on a patch to devstack-gate
which sets the right vars there.

This patch also adds a few comments explaining the peculiar test setup,
and a work-around for libvirt bug
  https://bugs.launchpad.net/ubuntu/+source/libvirt/+bug/1248025

Change-Id: Ibf2b2b0d2ae491793f291562d63299562bd83ebe