Drive puppet from the master
We'd like to be able to control sequencing of how and when puppet
runs across our machines. Currently, it's just a set of agents
that run kinda whenever they run. At times they hang and we don't
know about it. Also, cross-server sequencing is impossible to
achieve.

Change the operation away from agents running on the machine as
daemons, and instead use puppet-kick from the master to each machine.

Change-Id: I76e41e63c6d0825e8735c484ba4580d545515e43