Move channel reg requirement to top of IRC docs
We require channels to be registered and have the openstackinfra
account added to the access list before we'll add infra-managed
bots.

Moving this documentation to the beginning of the document and
giving them their own section in case we wish to have other
requirements in the future.

Also add link on stackforge.rst back to the IRC page.

Change-Id: Ic92f3935de19c4afef23f51257a189f2860796c2