Add rstcheck to the qa-specs repo
This commit adds a new job for the qa-specs repo to run a quick sanity
check against proposed changes. One note is that the run_bash8.sh slave
script is reused because to run rstcheck.py the run_test.sh in the root
dir of qa-specs is run. This is the same way bash8 is invoked so instead
of duplicating the slave script the bash8 one is used.

This depends on qa-specs change I48352a1d79f38df936f424bf9290430c0258d79d

Change-Id: I3cb2155cfa4d77e95a2c333867bab90d8f12fd94