make neutron not voting in the gate
neutron is now dealing with a deadlock architectural issue for 8
days. It's now created a 17hr backup in the gate during the
freeze period, substantially disrupting the ability for other
projects to land release critical bugs.

We should make it non voting until the deadlock can be resolved,
then we can bring this back. Because the -pg jobs are still enabled
on neutron this means neutron itself will still be self gating,
just no one else will gate on it.

Change-Id: Icc578bf38d54fe26efb8e6d3f4547c40a02e1575