propose_translation_update: Handle .tx/config in git
Some packages have .tx/config in git. The tx commands update
.tx/config if new localization files are added to the repository.
If .tx/config changes, "git review" will fail with a
message stating that there are uncommitted files in the tree.

Since .tx/config reflects the localization setup of the repository,
we should commit it as part of a change set if it changes.

Therefore, always add the file .tx/config if it is under git control.

Change-Id: I235877fb568c0365cc822c265d401f0cf4dfdbf5