propose_translation_update: Handle .tx/config in git
If .tx/config is in git and gets modified by the script, we need to add
it to our review request - otherwise "git review" will fail with a
message stating that there are uncommitted files in the tree.

Therefore, always add the file .tx/config if it is under git control.

Change-Id: I235877fb568c0365cc822c265d401f0cf4dfdbf5