pbx: don't install asterisk-sounds-extra-en-g722
This package appears to have gone missing from the repo we use.  Just
don't install it for now.  It's not a big deal.  It just means we only
have the 8 kHz versions of these files for now instead of the 16 kHz
versions for callers using g722.

I reported the issue upstream to the asterisk-dev mailing list.

Change-Id: I8a9117d3458a8f219b83f9e6b711f71407f5ee97