Work around for single hiera data dir
This moves the hiera data dir to /etc/puppet and drills into the
allows for two hiera data directories. One is the top secret data
currently in hiera. The other is $environment/config/hiera.

This enables data-like things that are not secret to be pulled out of
manifests and put into hiera. It also allows not-exactly secret lists of
things to come out of secret-hiera and go into public-hiera.

Examples are the $sysadmins variable that gets passed around everywhere.
It could simply be a hiera('sysadmins') lookup wherever we need it.

Another example is the list of jenkins plugins to install.

Change-Id: I962a325ca21d4a20432a0a1b0c16a85f6b1bb8de