Add --force to transifex pull
As noted in the mailing list:
http://lists.openstack.org/pipermail/openstack-i18n/2014-March/000479.html

Transifex pull commands were failing due to bad timestamp processing
in the transifex client.

This patch fixes the problems by forcing pulls, even if timestamps
are not aligned as expected - meaning that it will always get the
latest version of the file.

This results in slightly slower runtime, but will mean it is always
accurate.

Change-Id: Ia254cc381378998ae3ecec7950c8409a1a6400b8
Partial-Bug: 1299349