Limit Tempest tests in virutal Ironic devstack job
Build a big ugly regex and pass as DEVSTACK_GATE_TEMPEST_REGEX to
devstack-vm-gate.sh causing testr to omit Tempest tests that stress
features not supported by Ironic.

Also updates the job to pull the newer libvirt requirement from
cloud-archive:icehouse.  This allows us to drop the workaround for a RAX Xen
issue as well as get some early testing of the libvirt version to be shipped in
14.04.

Change-Id: I4fbbb195997afb662eb6be1c922d8adbacee749e