add a build-timeout macro
The 'BUILD_TIMEOUT' env variable references the build timeout value.  This
change adds a build-timeout macro to help simplify jenkins jobs.  The first
job to use the macro will be the jenkin-plugin-jobs.

This change requires the Jenkins build-timeout plugin > ver 1.13  It not work
with older versions of the build timeout plugin.  For the record if this
setting is used with an older version of the plugin (i tested with 1.10)
it will  set the timeout to 0 and abort builds immediately.

Partial-Bug: #1287883
Change-Id: I7dbe6e506d4f4045a6c509022e596641d5cf2ffd