update the jjb timeout global defaults and add a build-timeout macro
This change adds a timeout environment variable to the default jjb timeout
wrapper so that any job which uses this wrapper will get the variable
by default.  The 'BUILD_TIMEOUT' env variable references the build timeout
value.  This change also adds a build-timeout macro to help simplify
jenkins jobs.  The first job to use the macro will be the jenkin-plugin-jobs.

Partial-Bug: #1287883
Change-Id: I7dbe6e506d4f4045a6c509022e596641d5cf2ffd