Fix milestone-proposed requirements update
milestone-proposed was checked out in a detached state, which prevents
git-review from successfully proposing changes to it. It works with
master because master is checked out by default on the initial clone.

This commit makes sure the branch is checked out with proper origin
tracking before proceeding with changes.

Fixes bug #1302490

Change-Id: Iff4f653d2d68fd36bb076ed8ae9fa5b0bb03edda