Make ATC lists sort E-mails with preferred first
* tools/atc/email-stats.py: Adjust the order in which Gerrit E-mail
addresses are returned so that the one listed as the preferred
address is sorted to the beginning of the set. Also move procedural
code out of the global scope to make this file safely importable,
and clean up a couple of unused variables flake8 noticed in the
newly importable main codeblock.

Change-Id: I95b5817ada58f4059df4f5c0b406183f1c1e5b74