Move xenapi-os-testing to stackforge.
This repo contains scripts and supporting functionality for the XenServer CI
environment.  We want this to be under gerrit's control so we can ensure changes
do not break the XenServer CI, and so that others can propose changes primarily
to the tempest exclusion list.

Tests are automatically run by the xenserver 3rd party CI on this repository.

Change-Id: I7e9d0e8b917ca39f20f41c9758bffc4cecd9ec93