WIP: Use ansible instead of direct ssh calls
For sake of argument, this is what a python program version can
look like using ansible - and also sets the root account on puppetmaster
up for doing more interesting things.

Please note - I do not think this will work as is - as I have not tested
the intersection of how ansible may or may not choose to run the listed
command with how we're limiting command lists in the key.

Change-Id: I53112bd1f61d94c0521a32016c8a47c8cf9e50f7