Create infra-manual project
The idea is that it will be a three-part user's manual for the
OpenStack project infrastructure:

  * Developer's User Manual
  * Core Reviewer's User Manual
  * Project Driver's User Manual

It will contain information on how to work with git, git-review,
gerrit, zuul, jenkins, etc., all targeting those three kinds of
users.  It should completely replace this document:

  https://wiki.openstack.org/wiki/GerritJenkinsGit

And much of the content in:

  https://wiki.openstack.org/wiki/GerritWorkflow

As well as these parts of the current infra docs:

  http://ci.openstack.org/stackforge.html
  http://ci.openstack.org/test-infra-requirements.html

It will have its own core group that will include infra-core but
can easily be expanded to anyone who understands the infrastructure
and current practices well enough to review documentation changes
(which is quite a lot of developers).

Change-Id: I371cce0622c1ae99e9cdfa313b36ccf0bb256773