Stop trying to start gerrit
Because we don't control _when_ puppet runs, and because restarting
services can cause downtime, we tend to avoid having gerrit manage
their service state for things like restarts. Gerrit is one such
service - restarting it is costly - and things, such as the current
puppet bug which makes puppet think that the config has change every
time puppet is run, can cause puppet to think that it should restart
gerrit every 15 minutes. That's not a good thing.

Additionally, the current code for starting it does not work - we've
been getting spam in our logs for months about it.

To resolve, remove the non-working puppet service control line, and
leave things the way we want them, which is with puppet laying down
the configs and us controlling when and if we want gerrit to restart.

Change-Id: I3bcaaed07bcf6239e053789b9b6241cbf652e7d4