temporary disable gerrit tests
Need to temporarily disable gerrit tests because they seem to fail only
when run on HP slaves.  The tests do pass on rax slaves.  The failing tests
are blocking gerrit changes from merging.

This change shows tests passing on rax and failing on HP:
https://review.openstack.org/#/c/89699

This change wants to get merged but can't because tests keep running
on HP slaves:
https://review.openstack.org/#/c/89698

Change-Id: Iead7a7ba3934abde3ff647357694d7d3e223c4df