Add yum.openstack.org lightweight Fedora 20 mirror
Adds a new openstack_project manifest to create a lightweight
F20 mirror containing x86_64 and i386 testing/updates/release RPMs.

This will be useful for TripleO where we do lots of disk
image building and having a reliable mirror is critical.

The plan is to run this mirror in the public cloud and use
local Squid caches to avoid repeated downloads. We may also
choose to host individual mirrors in each TripleO CI cloud.

Tested on a Centos 6.5 host machine.

Change-Id: If47acc1e268a38b3bf1c06b2520472b9b089bff0