Skip running update-grub if it doesn't exist
update-grub doesn't exist on fedora so don't attempt to run it, the sed
command editing /etc/default/grub doesn't match GRUB_CMDLINE_LINUX="" so
isn't adding the mem=8G parameter anyways. We need to circle back to
this in order to find an appropriate alternative that will restrict
memory on fedora.

Change-Id: I21ce382a0023d674a4a2e6afecc9bc678791eac9