remove pypy from python jobs
It's now been nearly 8 weeks since we stopped voting on these jobs
because of the setup tools issue. There doesn't currently seem to
be anyone working on the issue to close it out across all projects.

2 months of not running a job seems like a reasonable threshold to
stop consuming nodes on unpassable execution.

This removes it from the core gate jobs where we see a high failure
rate, and makes pypy voting again for other jobs.

Change-Id: Ic35f75799a71abc9f89f79bd1128cf131a64283b