TripleO: use the common-venv element by default.
Updates all the TripleO CI jobs so that we include the common-venv
element in all of our diskimage builds by default.
The primary motivation for this change is that it should
help speed up CI (quite a bit for the overcloud) in that
it avoids recompiling the same python libraries in
multiple venvs.

Given that OpenStack now maintains a common 'requirements'
repo concerns about breakage should be minimized and this
may even help catch some backwards compatability issues
in various client libraries ahead of time as well (once
we gate on them).

Change-Id: Iac6156991ca37b6b42dc6d807ec6a7f5ed3d8bb6