Ensure jenkins user has useful secure_path
I am seeing calls to 'useradd' fail as "not found" in the devstack
setup scripts on Fedora 20.  /usr/sbin should be in the path for the
user by default; this must be being overridden by Jenkins although I
can't see where.

Rather than relying on external influences on PATH, this changes
ensures that the jenkins user always has a sane path to work with
under sudo.

Change-Id: I3217f20910027c90544819aac2328016621640e2