Create telemetry-specs git repo
This will serve as a repository of approved blueprint
specifications for the Telemetry program, from Juno
onwards.

The gerrit review workflow for this repo will act as
the blueprint oversight and approval process for the
project.

Once landed, the approved version of a blueprint spec
is linked to the corresponding launchpad blueprint via
the specification URL being set in launchpad to:

  http://git.openstack.org/cgit/openstack/telemetry-specs/tree/specs/juno/${BLUEPRINT_NAME}.rst

Change-Id: I8a31a16dab6e16ed4b14558c6b00ff03b2f84132