add tempest jobs to dib and os-* tools
There is currently a devstack patch series starting at
Ia911cdee86f5b2e2ba1557e5aa8bf392b92ef555 which makes dib and os-*
available to ironic and heat as part of devstack, as something to
gate on instead of downloading static images. As such, we need
to put these into the co-gating mechanism so that we ensure these
tools don't break us when updating.

Change-Id: I7f02ff99fde0b8dfea235e79b9a874c6e89bba77