Create common translation functions
Currently there's a lot of duplicated code between the translation
scripts. This is an initial step in creation of a common library that
avoids the duplication.

Move the setup of git and transifex in files
propose_translation_update.sh and upstream_translation_update.sh into
two new functions in the new file common_translation_update.sh and
source the library and use the files.

Change-Id: I7831909e780e6f683bbc7c2fd67f46232d9d1fdc