Add support for disk-image-builder in nodepool
In order to use disk-image-builder to build nodepool images, we need
a few things. First, we need additional debs to be installed for
disk-image-builder to work. Then we need the image elements themselves
to describe the images we want to create. Also, a few helper scripts
for locally working with the disk images have been included.

Diff to 41

 - modify image-dir in nodepool.yaml.erb to match directory provided
   by [1]
 - drop blank change in manifests/site.pp

Tested with build-image.sh script; devstack-gate-precise.qcow2 created
correctly

[1] https://review.openstack.org/#/c/102583/

Change-Id: I539743147341dc5b387d103fecd3e9ff0f01fdfc