Add XStatic-* projects with packaged static files for Horizon
Horizon will no longer bundle all the static JavaScript and style files
in its repository. It will instead depend on XStatic packages with those
files included in them.

Every library is a separate XStatic package, so that it can be tracked
and versioned separately and possibly shared with other projects that
also use XStatic.

This has been discussed on a design session on the summit:
https://etherpad.openstack.org/p/juno-summit-horizon-static-files

Change-Id: Ie087c173b531faab4ae6a55f4a05e7ac3bd990c9
Implements: https://blueprints.launchpad.net/horizon/+spec/remove-javascript-bundling