Add button that shows/hides CI comments in Gerrit
There are many CIs which automatically post comments in Gerrit and there
is no way to filter those out from the normal comments posted by
reviewers.

This patch adds a new "Toggle CI" button to the Gerrit interface which
shows/hides CI comments. The button is displayed only on review pages
and is hidden everywhere else.

Change-Id: I6ffca12d6314c63c0f1ab97d2c98a84a6ac790c9
Closes-Bug: 1282019