Remove pip module from config
This commit removes the pip module from the
modules that are embedded in the config project.

This change is part of a larger set of changes
to decouple core gating components so that
a user can deploy a partial set of the openstack-infra
(for my use case, the minimum set of components required
to set up federated gating based on the same set of modules)

Pip was selected because it is a dependency of jenkins and zuul
(which I also intend to split out)

NOTE: This patch does not work yet, b/c it assumes that v0.0.1 of
pip has been released to the forge (which it has not!). I have a version
of pip that is pretty close to being ready at: bodepd/puppet-pip, the
intention is that this needs to be moved to openstack-infra/puppet-pip.

Change-Id: If73e8a2d86e47f041770b4b906eddfa1f4f18434