Remove unused pypi jobs from projects
Some projects have pypi-jobs listed in the jjb config
but do not actually use them (defined as not having
a package on pypi using the same name as the project
here, which may be overly aggressive at pruning).

This change adds a script to find such projects, and
then removes several of them from the projects list.

Change-Id: I2c125c209b06f7148e2a144e5b8537ac8e415b8c