Move large-ops-testing to silent queue for further testing
The large-ops-testing jobs (nova-network and neutron) is being used to
find a new suitable large-ops number, large-ops currently uses 50, while
large-ops-testing is using 100.

The large-ops number was originally 150, but that had to be dropped due
to performance variance in the gate. Because of this, running
large-ops-testing a few times in the experimental queue was good enough
to see if a number would work at all (175 didn't work because of
ceilometer), but its not enough data to see if the new number is
reliable enough for the gate.

Keep large-ops-testing on nova only, as this should result in a good
amount of data that can be analyzed via graphite while not adding an
additional job to every patch.

If this job works reliably for a week or two, the next step is to turn
it off again and bump up the LARGE_OPS_NUMBER in the regular large-ops
tests to 100.

Change-Id: Ib0dd2931efdb4b561570a6fef04f93c2fdd017e9