New Chef cookbooks for client services, Sahara and Trove.
Added "cookbook-openstack-data" for supporting the Sahara project. The wiki page doesn't call it anything more than the Hadoop service, so it was decided to call it the 'openstack-data' cookbook.

Added "cookbook-openstack-database" for supporting the Trove project. There was an active discussion and voting around naming the repository, but since other OpenStack services do not include -service in their name, "openstack-database" was chosen. We should be able to keep this straight from the "database", "openstack-data" and the "openstack-ops-database" cookbooks

Added "cookbook-openstack-client" to serve the need for a cookbook containing Resources for interacting with OpenStack services (Cinder, Heat, Swift, etc.). Will mostly contain LWRPs for interacting via recipes and with Chef Metal.

Change-Id: Ifffe3c24617092a530f5e31741e24cff3154e04a