Add Stackforge project for training-guides
The OpenStack Training Guides team has been sheltering under
the loving arms of the Documentation team for the recent past.

It's now time to move them out into the big bad world, along
their path to success.

This change adds a project:

stackforge/openstack-training-guides

seeded with the training-guides subdirectory from openstack-manuals,
and attempts to configure the jenkins/zuul jobs so that the existing
docbook check, gate and post will also run on the new project.

Submitted on behalf of Sean Roberts.

Change-Id: I1aed9efa8ca3fd106872cec248cd7c4964fa9adf