add devstack-vagrant project to openstack-dev
devstack-vagrant is a set of vagrant tooling around devstack to
make it super easy to bring up devstack locally on any machine
that supports vagrant. It includes a few top level config options
to bring up some custom configs of devstack, and will be evolved
to make it easy to replicate very closely the configuration of
devstack-gate.

Change-Id: I24390f5fbd582b8d7d268c6460ac24479c63719f