Fixed config of manila's jobs
Env vars from separate script were exported only
for child process and were not available for other subprocesses.
https://jenkins07.openstack.org/job/gate-manila-tempest-dsvm-neutron/122/console

Set all external scripts to be called in parent process.
Because there is no need to use subprocesses.

Call of setup.sh is intended to be in pre_test_hook:
https://review.openstack.org/#/c/97003/

Also added 'export DEVSTACK_GATE_NEUTRON=1' to the config,
because name of job supposes using of neutron, so enable it
by default.

Change-Id: Ia211526d63c172b5c9f700d59c143ad78694348b