Add in new check-non-binding pipeline
This pipeline is for jobs that don't want to be weighing in on the
check vote but also don't want to be overlooked by being non-voting.

This comes about due to the use case of ironic needing to vote on
nova without dirtying the Jenkin's vote or being missed
http://lists.openstack.org/pipermail/openstack-dev/2014-May/036174.html

This depends on a change to zul to support gerrit-user in the reporter
(I9e503e3bccd7b28f6a1250c615b13b124146c62a)

It also requires a user 'non-binding-ci' to be created with the same SSH key as 'jenkins'.

Change-Id: I10c8e47610b0d43cdb3309560334360a5e0b7321