Extend graphite storage schema configuration
Configures graphite with a templated storage-schemas.conf.  This
setup defaults to applying the O.O settings for this file, but also
allows that default to be unset, and new entries applied like so:

  graphite::storage { 'default_1min_for_1day':
    pattern    => '.*'
    retentions => '60s:1d'
  }

This satisfies a condition downstream where the configuration of
graphite results in an 'out of disk space' condition, and where
a smaller retention set is desired.

Change-Id: I93c9899d7f789c855286fb1b79c592492575badc