upgrade jenkins build-timeout plugin due to a performance bug
There was a sever performance bug in version 1.13 :
https://issues.jenkins-ci.org/browse/JENKINS-23012

I have tested version 1.14 on jenkins_dev.o.o and have verified that I could
not reproduce the performance problem reported in ver 1.13.

I think it's safe to use this version because plugin does not enable the
log watching feature by default and none of our projects currently use
that feature.

Change-Id: I82f60ebe352d33e4664fdbc1ff1aa876527cb23a