Move variables in manifests/site.pp to top of file
Put the variables defined in manifests/site.pp at the top of the file so
that they are in a known location after running csplit in test.sh and we
can prepend them to the puppet-apply top files.

A better solution would be to move this data into hiera, but this is not
sensitive data, so a move to hiera should wait until we a solution for a
public hiera data repo separate from the private hiera data.

Change-Id: I509a8266462dfdf53e1727938e4fb043241166b6