Modularized StoryBoard Module
In order to get the puppet module for storyboard up to a level where we can
publish it to puppetforge, I did some work on it to create separate modules
which can be used by anyone to install storyboard.

- API and Webclient are now installed via storyboard::application, which
assumes that you can provide the DB connection criteria.
- storyboard::cert is now a separate class, which accepts either files or strings,
which generates the SSL certificate and chain files for storyboard.
- storyboard::params is our dependency checker.
- storyboard::init will install a standalone, entirely self-contained instance
of storyboard.
- Added various puppet module files necessary for eventual deployment
to puppetforge.
- Added README.md documentation for later puppetforge addition.

This patch also includes a new module: example42-puppi, which is a series
of convenience utilities useful for deployment. For example, puppi::netinstall
(used here) will fetch tarballs and zip files and extract them into a provided
directory. It also contains changes to the storyboard configuration for the
new refresh token support patch in #94363

Change-Id: I6ab8c24b308df38774fc0694d218dcb5022cd899