Promote devstack/f20 job to check job
The experimental devstack/f20 job has been quite stable, and [1] will
provide redundancy for the test.

This change moves the existing devstack/f20 it to a non-voting
devstack check job and renames it to remove the "experimental" and
"devstack" parts of the name; this test will be useful in other areas
such as devstack-gate script tests, so it's not specific to devstack
changes.

[1] https://review.openstack.org/#/c/98024/

Change-Id: I7670ee2ec805b46cf68294109e2018569db843c8