Fix nodepool instance hostnames
In some clouds the nodepool instance will not be able to resolve it's own
hostname. And in some clouds the hostname will not be the value nodepool
uses for the instance name. So in some case we need to set our hostname
or make it resolve. Since having a resolvable hostname is best practice,
let's apply the fix to the instance used to create the image as well as
the instances booted from the created image.

Change-Id: I1061bbb2c89bc973f77ba9d43c1de11ae178f796