Run stackalytics in infra
Stackalytics has become the de-facto standard stats reporting tool
for the OpenStack community. As such, we should run it as part of
the projects infrastructure. Once this is up and running, we can ask
Mirantis to point the CNAME to stackalytics.openstack.org and/or
transfer control of the domain to the foundation.

Change-Id: I294c0aa232c67ed50d61a7c4cc4a6b65d84ed162