Adjust nodepool limits for rax providers
Adjust nodepool limits for rax-dfw, rax-iad and rax-ord to reflect
our current reality. Our primary quota constraints there are
maxTotalInstances and maxTotalRAMSize, the latter needing to be
divided by 8192 to get an upper bound on possible node count. Take
whichever is lower in each region, subtract the number of images
configured for it to make room for image update template instances
while running at capacity, and knock off two more for breathing
room.

The drastic reduction in max-servers for rax-iad is due to a recent
and so far unexplained drop in maxTotalRAMSize there, which is being
investigated separately and can be readjusted upward once the cause
is addressed.

Change-Id: Iec601ed87bae9a048525ebcde37deb373b688f4d