Don't notify nova's IRC room on patch creation
With all the new patches (and revisions) being submitted to gerrit all
the time it has become hard to keep up with nova's IRC room when large
chunks of the logs are just the openstackgerrit bot.

Also with such strong reliance on automated testing, folks commonly
won't look at a patch that doesn't have test results yet, so the
creation of a new patch is rarely an actionable event for most users.

Change-Id: I70b3413029eebbe84896a7b8ff61e3035d8d1844