do not co-gate oslotest with the projects that include it
A side effect of fully symmetric testing of the olsotest module is
that it will end up joining the world into a single gating queue
due to transitive dependencies. It already put ironic in line with
the rest of the gate because the following cross testing:

  ironic <-> dib <-> oslotest <-> oslo.messaging <-> all of openstack

As stackforge projects included oslotest, we might have gotten
similar things.

Here is the proposed solution to get us out of this:

 * remove oslotest tests from the gate queue of projects that use it
 * continue to run oslotest tests on check queues for those projects
   (this is basically 'gate lite', as clean check still makes them
    have to pass these tests)
 * leave oslotest gate tests on oslotest. This handles the case where
   a series of olsotest patches merge conflict badly, but can't get
   caught until the gate.

This means there is an exposure where "just the right" set of patch
lands in a project which break oslotest usage could cause a temporary
wedge of that project + oslotest. However I think that's a localized
enough problem we should take that risk and manage it instead of
the current monolythic pipeline problem.

Change-Id: I371ddf7c1de42d82773468064b5706ae5c36ad8b