Don't run large-ops test on stable/havana branches
The large-ops test is there to prevent massive performance regressions,
and we are ratcheting up this test over time. We don't want to have to
maintain a separate large-ops job for stable/havana especially since
stable releases only accept 'safe' backports from master. So if a patch
is being considered for backport it already has passed master large-ops
and a human reviewer evaluating the patch explicitly for how safe it is
to backport.

The new large-ops number broke this test for stable/havana, so turn off
the job on stable/havana but keep it on stable/icehouse as its still
working there.

Closes-Bug: #1329042

Change-Id: I5fb64f8537dfffca79170000310a02353b7d1b0e