Don't run large-ops test on stable branches
The large-ops test is there to prevent massive performance regressions,
and we are ratcheting up this test over time. We don't want to have to
maintain a separate large-ops job for stable releases especially since
stable releases only accept 'safe' backports from master. So if a patch
is being considered for backport it already has passed master large-ops
and a human reviewer evaluating the patch explicitly for how safe it is
to backport.

Change-Id: I5fb64f8537dfffca79170000310a02353b7d1b0e