Add manuals-jobs for training-guides
The training-guides have their own repo now, but their manuals are no
longer published to docs.openstack.org.

The current infra-config contains only a partial solution. There is a
project in projects.yaml, but no matching job in manuals-jobs.yaml.

This patch tries to fix that problem by adding a job for training-guides
in manuals-jobs.yaml.

Closes-Bug: 1329567
Change-Id: Ieaf238efdd2d955bb1c5c71dcc7cd1ddc4773edf